scenario exec - rename, reorg, cleanup
