+++ framework/scenario/src/facade/world_tx.rs
@@ -1,9 +1,9 @@
-#![allow(unused)] // TEMP
-
+mod scenario_check_state;
 mod scenario_exec_call;
 mod scenario_exec_deploy;
 mod scenario_query_call;
 mod scenario_rh_impl;
+mod scenario_set_state;
 mod scenario_tx_env;
 
 pub use scenario_exec_call::ScenarioEnvExec;
